requirements.txt
openai
python-dotenv
streamlit
pydub
pandas
pycaret
